refactor: 패키지이름을 type -> formtype으로 변경
