+++ survey/survey-application/src/main/java/me/nalab/survey/application/port/in/web/findfeedback/formtype/FeedbackFindByTypeUseCase.java
@@ -1,4 +1,4 @@
-package me.nalab.survey.application.port.in.web.findfeedback.type;
+package me.nalab.survey.application.port.in.web.findfeedback.formtype;
 
 import java.util.List;
