+++ survey/survey-application/src/main/java/module-info.java
@@ -34,7 +34,7 @@ module luffy.survey.application.main {
 
 	exports me.nalab.survey.application.port.out.persistence.findtarget;
 	exports me.nalab.survey.application.port.in.web.findtarget;
-	exports me.nalab.survey.application.port.in.web.findfeedback.type;
+	exports me.nalab.survey.application.port.in.web.findfeedback.formtype;
 
 	requires luffy.survey.domain.main;
 	requires luffy.core.id.generator.id.generator.starter.main;
